11/29/21 -- Adam -- Moved objects into a class file trying to fix memory problem with the instances -- 1 hour
