+++ VentureForge/VentureForge.csproj
@@ -55,6 +55,7 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
+    <Compile Include="Memory.cs" />
     <Compile Include="CreateModule.cs">
       <DependentUpon>CreateModule.xaml</DependentUpon>
     </Compile>
